delete_moodle_entities/requirements.txt
python-dotenv
requests
munch
structlog
progressbar2
